+++ src/adminsite/views/__init__.py
@@ -1,4 +1,5 @@
+from adminsite.views.inline import Inline, InlineRow
 from adminsite.views.model_view import ModelView
 from adminsite.views.registry import ViewRegistry
 
-__all__ = ["ModelView", "ViewRegistry"]
+__all__ = ["Inline", "InlineRow", "ModelView", "ViewRegistry"]